+++ tests/MsGraphAdminTest.php
@@ -1,9 +1,23 @@
 <?php
 
+use Dcblogdev\MsGraph\Models\MsGraphToken;
 use Dcblogdev\MsGraph\MsGraphAdmin;
 
-test('can initalise', function () {
-    $msGraphAdmin = new MsGraphAdmin();
+beforeEach(function () {
+    $this->msGraphAdminMock = Mockery::mock(MsGraphAdmin::class);
+});
 
-    $this->assertInstanceOf(MsGraphAdmin::class, $msGraphAdmin);
+test('can initalise', function () {
+    $this->assertInstanceOf(MsGraphAdmin::class, new MsGraphAdmin());
+});
+
+test('can refresh token', function () {
+    MsGraphToken::create([
+        'user_id'      => null,
+        'access_token' => 'ghgh4h22',
+        'expires'      => strtotime('+1 day'),
+    ]);
+
+    $this->artisan('msgraphadmin:keep-alive')
+        ->expectsOutput('connected');
 });